Update user_stories.md
Hi Kylie, 
For our Milestone 3 Progress Report this repository needs to be neatened up. I have updated the User Stories table to include details from the User Stories Excel spreadsheet (G drive). I have also finalised wording on some of the outstanding US and merged some of the US based on comments in the spreadsheet. I did this a copy and paste of the whole table, so you won't be able to see the individual line item changed. I've added additional comments to the Excel spreadsheet to outline what I did. 
Let me know if there are any issues or items that need clarifying. 
Cheers, Miranda